backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
torch
diffusers
transformers
accelerate
safetensors
python-dotenv
peft

